Refine PPO training layout and documentation
